Add internal load method. Fix HomeWorkspaceModel constructor.
