+++ src/DynamoCore/Models/DynamoModel.cs
@@ -20,11 +20,13 @@ using Dynamo.Scheduler;
 using Dynamo.Search;
 using Dynamo.Search.SearchElements;
 using Dynamo.Selection;
+using Dynamo.Serialization;
 using Dynamo.Updates;
 using Dynamo.Utilities;
 using DynamoServices;
 using DynamoUnits;
 using Greg;
+using Newtonsoft.Json;
 using ProtoCore;
 using ProtoCore.Runtime;
 using System;
@@ -1383,6 +1385,70 @@ namespace Dynamo.Models
                 newWorkspace.EvaluationCompleted -= OnEvaluationCompleted;
                 newWorkspace.RefreshCompleted -= OnRefreshCompleted;
             };
+        }
+
+        internal void Load(string workspaceJson)
+        {
+            var settings = new JsonSerializerSettings
+            {
+                Error = (sender, args) =>
+                {
+                    args.ErrorContext.Handled = true;
+                    Console.WriteLine(args.ErrorContext.Error);
+                },
+                ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
+                TypeNameHandling = TypeNameHandling.Auto,
+                Formatting = Newtonsoft.Json.Formatting.Indented,
+                PreserveReferencesHandling = PreserveReferencesHandling.Objects,
+                Converters = new List<JsonConverter>{
+                    new FunctionDescriptorConverter(LibraryServices),
+                    new CodeBlockNodeConverter(LibraryServices),
+                    new ConnectorConverter(), new AnnotationConverter()
+                },
+                ReferenceResolverProvider = () => { return new IdReferenceResolver(); }
+            };
+
+            var ws = JsonConvert.DeserializeObject<HomeWorkspaceModel>(workspaceJson, settings);
+
+            // TODO: #4258
+            // The logic to remove all other home workspaces from the model
+            // was moved from the ViewModel. When #4258 is implemented, we will need to
+            // remove this step.
+            var currentHomeSpaces = Workspaces.OfType<HomeWorkspaceModel>().ToList();
+            if (currentHomeSpaces.Any())
+            {
+                // If the workspace we're opening is a home workspace,
+                // then remove all the other home workspaces. Otherwise,
+                // Remove all but the first home workspace.
+                var end = ws is HomeWorkspaceModel ? 0 : 1;
+
+                for (var i = currentHomeSpaces.Count - 1; i >= end; i--)
+                {
+                    RemoveWorkspace(currentHomeSpaces[i]);
+                }
+            }
+
+            AddWorkspace(ws);
+
+            // TODO: #4258
+            // The following logic to start periodic evaluation will need to be moved
+            // inside of the HomeWorkspaceModel's constructor.  It cannot be there today
+            // as it causes an immediate crash due to the above ResetEngine call.
+            var hws = ws as HomeWorkspaceModel;
+            if (hws != null)
+            {
+                // TODO: #4258
+                // Remove this ResetEngine call when multiple home workspaces is supported.
+                // This call formerly lived in DynamoViewModel
+                ResetEngine();
+
+                if (hws.RunSettings.RunType == RunType.Periodic)
+                {
+                    hws.StartPeriodicEvaluation();
+                }
+            }
+
+            CurrentWorkspace = ws;
         }
 
         #endregion
