+++ src/DynamoCore/Serialization/SerializationConverters.cs
@@ -1,0 +1,239 @@
+﻿using Dynamo.Engine;
+using Dynamo.Graph.Annotations;
+using Dynamo.Graph.Connectors;
+using Dynamo.Graph.Nodes;
+using Dynamo.Graph.Notes;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Linq;
+using Newtonsoft.Json.Serialization;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Dynamo.Serialization
+{
+
+    /// <summary>
+    /// The AnnotationConverter is used to serialize and deserialize AnnotationModels.
+    /// The SelectedModels property on AnnotationModel is a list of references
+    /// to ModelBase objects. During serialization we want to refer to these objects
+    /// by their ids. During deserialization, we use the ReferenceResolver to
+    /// find the correct ModelBase instances to reference.
+    /// </summary>
+    public class AnnotationConverter : JsonConverter
+    {
+        public override bool CanConvert(Type objectType)
+        {
+            return objectType == typeof(AnnotationModel);
+        }
+
+        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        {
+            var obj = JObject.Load(reader);
+            var title = obj["Title"].Value<string>();
+            var guid = Guid.Parse(obj["Uuid"].Value<string>());
+
+            // This is a collection of string Guids, which
+            // should be accessible in the ReferenceResolver.
+            var models = obj["SelectedModels"].Values<JValue>();
+
+            var existing = models.Select(m => serializer.ReferenceResolver.ResolveReference(serializer.Context, m.Value<string>()));
+
+            var nodes = existing.Where(m => typeof(NodeModel).IsAssignableFrom(m.GetType())).Cast<NodeModel>();
+            var notes = existing.Where(m => typeof(NoteModel).IsAssignableFrom(m.GetType())).Cast<NoteModel>();
+
+            var anno = new AnnotationModel(nodes, notes);
+            anno.AnnotationText = title;
+            anno.GUID = guid;
+
+            return anno;
+        }
+
+        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        {
+            var anno = (AnnotationModel)value;
+
+            writer.WriteStartObject();
+
+            writer.WritePropertyName("Title");
+            writer.WriteValue(anno.AnnotationText);
+            writer.WritePropertyName("SelectedModels");
+            writer.WriteStartArray();
+            foreach (var m in anno.SelectedModels)
+            {
+                writer.WriteValue(m.GUID.ToString());
+            }
+            writer.WriteEndArray();
+            writer.WritePropertyName("Uuid");
+            writer.WriteValue(anno.GUID.ToString());
+
+            writer.WriteEndObject();
+        }
+    }
+
+    /// <summary>
+    /// The FunctionDescriptorConverter is responsible for deserializing
+    /// and serializing the FunctionDescription property on DSFunction. 
+    /// Because a lookup in LibraryServices is required during deserialization,
+    /// we use this converter to find the correct FunctionDescriptor, and
+    /// call a node constructor which constructs a ZeroTouchNodeController
+    /// using the FunctionDescriptor.
+    /// </summary>
+    public class FunctionDescriptorConverter : JsonConverter
+    {
+        /// <summary>
+        /// A reference to an instance of the LibraryServices class.
+        /// This is required to properly setup the function given
+        /// the assembly and function name.
+        /// </summary>
+        private LibraryServices libraryServices;
+
+        public override bool CanConvert(Type objectType)
+        {
+            return objectType.IsAssignableFrom(typeof(FunctionDescriptor));
+        }
+
+        public FunctionDescriptorConverter(LibraryServices libraryServices)
+        {
+            this.libraryServices = libraryServices;
+        }
+
+        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        {
+            var jObject = JObject.Load(reader);
+
+            var asm = jObject["Assembly"].Value<string>();
+            var mangledName = jObject["Name"].Value<string>();
+
+            return string.IsNullOrEmpty(asm) ?
+                libraryServices.GetFunctionDescriptor(mangledName) :
+                libraryServices.GetFunctionDescriptor(asm, mangledName);
+        }
+
+        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        {
+            var fd = (FunctionDescriptor)value;
+            writer.WriteStartObject();
+            writer.WritePropertyName("Assembly");
+            writer.WriteValue(fd.IsBuiltIn ? "" : fd.Assembly);
+            writer.WritePropertyName("Name");
+            writer.WriteValue(fd.MangledName);
+            writer.WriteEndObject();
+        }
+    }
+
+    /// <summary>
+    /// The CodeBlockNodeConverter is used to serialize and deserialize CodeBlockNodeModels. 
+    /// CodeBlockNodeModel requires an instance of LibraryServices for construction.
+    /// </summary>
+    public class CodeBlockNodeConverter : CustomCreationConverter<CodeBlockNodeModel>
+    {
+        private LibraryServices libraryServices;
+
+        public CodeBlockNodeConverter(LibraryServices libraryServices)
+        {
+            this.libraryServices = libraryServices;
+        }
+
+        public override CodeBlockNodeModel Create(Type objectType)
+        {
+            return new CodeBlockNodeModel(libraryServices);
+        }
+    }
+
+    /// <summary>
+    /// The ConnectorConverter is used to serialize and deserialize ConnectorModels.
+    /// The Start and End of a ConnectorModel are references to PortModels, but
+    /// we want the serialized representation of a Connector to reference these 
+    /// ports by Id. This converter resolves the reference to the correct NodeModel
+    /// instance by id, and constructs the ConnectorModel.
+    /// </summary>
+    public class ConnectorConverter : JsonConverter
+    {
+        public override bool CanConvert(Type objectType)
+        {
+            return objectType == typeof(ConnectorModel);
+        }
+
+        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        {
+            var obj = JObject.Load(reader);
+            var startId = obj["Start"].Value<string>();
+            var endId = obj["End"].Value<string>();
+
+            var startPort = (PortModel)serializer.ReferenceResolver.ResolveReference(serializer.Context, startId);
+            var endPort = (PortModel)serializer.ReferenceResolver.ResolveReference(serializer.Context, endId);
+
+            var connectorId = Guid.Parse(obj["Uuid"].Value<string>());
+            return new ConnectorModel(startPort, endPort, connectorId);
+        }
+
+        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        {
+            var connector = (ConnectorModel)value;
+
+            writer.WriteStartObject();
+            writer.WritePropertyName("Start");
+            writer.WriteValue(connector.Start.GUID.ToString());
+            writer.WritePropertyName("End");
+            writer.WriteValue(connector.End.GUID.ToString());
+            writer.WritePropertyName("Uuid");
+            writer.WriteValue(connector.GUID.ToString());
+            writer.WriteEndObject();
+        }
+    }
+
+    /// <summary>
+    /// The IdReferenceResolver class allows us to use the Guid of
+    /// an object as the reference id during serialization.
+    /// </summary>
+    public class IdReferenceResolver : IReferenceResolver
+    {
+        private readonly IDictionary<Guid, object> models = new Dictionary<Guid, object>();
+
+        public void AddReference(object context, string reference, object value)
+        {
+            Guid id = new Guid(reference);
+            models[id] = value;
+        }
+
+        private static Guid GetGuidPropertyValue(object value)
+        {
+            // Use reflection to find the Guid or the GUID
+            // property on the object.
+
+            var pi = value.GetType().GetProperty("Guid");
+            if (pi == null)
+            {
+                pi = value.GetType().GetProperty("GUID");
+            }
+
+            var id = pi == null ? Guid.NewGuid() : (Guid)pi.GetValue(value);
+            return id;
+        }
+
+        public string GetReference(object context, object value)
+        {
+            models[GetGuidPropertyValue(value)] = value;
+
+            return GetGuidPropertyValue(value).ToString();
+        }
+
+        public bool IsReferenced(object context, object value)
+        {
+            var id = GetGuidPropertyValue(value);
+            return models.ContainsKey(id);
+        }
+
+        public object ResolveReference(object context, string reference)
+        {
+            var id = new Guid(reference);
+
+            object model;
+            models.TryGetValue(id, out model);
+
+            return model;
+        }
+    }
+}
